更新管理员homepage几个功能并更改文件名
